academy/airflow_project/requirements.txt
apache-airflow
pandas
sqlalchemy
pymysql
ipykernel
openpyxl